Note about fonts
I initially had issues with the highest block, that was not showing fine. It turned out that the font I had in the terminal - "Monaco" - was providing its own rendition of that block, without providing the intermediate ones. Switching to Menlo fixed it. I don't know if you want to "make names" in the README.md though, so I kept the message generic.