Move test files to tests/ subdirectory
For easy updates (via Git), I'd like to simply clone the upstream repo and put the working copy root dir into my PATH, so that the spark command becomes accessible without needing to copy it around. Unfortunately, the "test" command also is in the root, and eclipses the default /usr/bin/test (which usually isn't used as most shells provide a built-in test command, but for those few cases where the external one is used, it's a nasty surprise when the test command suddenly talks about a roundup command or even starts executing tests!)
As the test command is auxiliary, put it (and the spark-test.sh it wraps) into a tests subdirectory, and adapt the Travis configuration.